+++ dw1000/src/hl/receiving.rs
@@ -1,9 +1,4 @@
-use crate::{
-    configs::{BitRate, SfdSequence},
-    mac,
-    time::Instant,
-    Error, Ready, RxConfig, DW1000,
-};
+use crate::{mac, time::Instant, Error, Ready, DW1000};
 use byte::BytesExt as _;
 use core::convert::TryInto;
 use embedded_hal::spi::SpiDevice;
@@ -61,168 +56,7 @@ where
     SPI: SpiDevice,
     RECEIVING: Receiving,
 {
-    pub(super) fn start_receiving(&mut self, config: RxConfig) -> Result<(), Error<SPI>> {
-        // Really weird thing about double buffering I can't find anything about.
-        // When a message is received in double buffer mode that should be filtered out,
-        // the radio gives a really short fake interrupt.
-        // This messes up all the logic, so unless a solution can be found we simply don't support it.
-        if RECEIVING::DOUBLE_BUFFERED && config.frame_filtering {
-            return Err(Error::RxConfigFrameFilteringUnsupported);
-        }
-
-        // For unknown reasons, the DW1000 gets stuck in RX mode without ever
-        // receiving anything, after receiving one good frame. Reset the
-        // receiver to make sure its in a valid state before attempting to
-        // receive anything.
-        self.ll.pmsc_ctrl0().modify(
-            |_, w| w.softreset(0b1110), // reset receiver
-        )?;
-        self.ll.pmsc_ctrl0().modify(
-            |_, w| w.softreset(0b1111), // clear reset
-        )?;
-
-        // We're already resetting the receiver in the previous step, and that's
-        // good enough to make my example program that's both sending and
-        // receiving work very reliably over many hours (that's not to say it
-        // becomes unreliable after those hours, that's just when my test
-        // stopped). However, I've seen problems with an example program that
-        // only received, never sent, data. That got itself into some weird
-        // state where it couldn't receive anymore.
-        // I suspect that's because that example didn't have the following line
-        // of code, while the send/receive example had that line of code, being
-        // called from `send`.
-        // While I haven't, as of this writing, run any hours-long tests to
-        // confirm this does indeed fix the receive-only example, it seems
-        // (based on my eyeball-only measurements) that the RX/TX example is
-        // dropping fewer frames now.
-        self.force_idle(false)?;
-
-        self.ll.sys_cfg().modify(|_, w| {
-            w.ffen(config.frame_filtering as u8) // enable or disable frame filtering
-                .ffab(0b1) // receive beacon frames
-                .ffad(0b1) // receive data frames
-                .ffaa(0b1) // receive acknowledgement frames
-                .ffam(0b1) // receive MAC command frames
-                // Set the double buffering and auto re-enable
-                .dis_drxb(!RECEIVING::DOUBLE_BUFFERED as u8)
-                .rxautr(RECEIVING::AUTO_RX_REENABLE as u8)
-                // Set whether the receiver should look for 110kbps or 850/6800kbps messages
-                .rxm110k((config.bitrate == BitRate::Kbps110) as u8)
-        })?;
-
-        // Set PLLLDT bit in EC_CTRL. According to the documentation of the
-        // CLKPLL_LL bit in SYS_STATUS, this bit needs to be set to ensure the
-        // reliable operation of the CLKPLL_LL bit. Since I've seen that bit
-        // being set, I want to make sure I'm not just seeing crap.
-        self.ll.ec_ctrl().modify(|_, w| w.pllldt(0b1))?;
-
-        // Now that PLLLDT is set, clear all bits in SYS_STATUS that depend on
-        // it for reliable operation. After that is done, these bits should work
-        // reliably.
-        self.ll
-            .sys_status()
-            .write(|w| w.cplock(0b1).clkpll_ll(0b1))?;
-
-        // Apply the config
-        self.ll.chan_ctrl().modify(|_, w| {
-            w.tx_chan(config.channel as u8)
-                .rx_chan(config.channel as u8)
-                .dwsfd(
-                    (config.sfd_sequence == SfdSequence::Decawave
-                        || config.sfd_sequence == SfdSequence::DecawaveAlt)
-                        as u8,
-                )
-                .rxprf(config.pulse_repetition_frequency as u8)
-                .tnssfd(
-                    (config.sfd_sequence == SfdSequence::User
-                        || config.sfd_sequence == SfdSequence::DecawaveAlt)
-                        as u8,
-                )
-                .rnssfd(
-                    (config.sfd_sequence == SfdSequence::User
-                        || config.sfd_sequence == SfdSequence::DecawaveAlt)
-                        as u8,
-                )
-                .tx_pcode(
-                    config
-                        .channel
-                        .get_recommended_preamble_code(config.pulse_repetition_frequency),
-                )
-                .rx_pcode(
-                    config
-                        .channel
-                        .get_recommended_preamble_code(config.pulse_repetition_frequency),
-                )
-        })?;
-
-        match config.sfd_sequence {
-            SfdSequence::IEEE => {} // IEEE has predefined sfd lengths and the register has no effect.
-            SfdSequence::Decawave => self.ll.sfd_length().write(|w| w.value(8))?, // This isn't entirely necessary as the Decawave8 settings in chan_ctrl already force it to 8
-            SfdSequence::DecawaveAlt => self.ll.sfd_length().write(|w| w.value(16))?, // Set to 16
-            SfdSequence::User => {} // Users are responsible for setting the lengths themselves
-        }
-
-        // Set general tuning
-        self.ll.drx_tune0b().write(|w| {
-            w.value(
-                config
-                    .bitrate
-                    .get_recommended_drx_tune0b(config.sfd_sequence),
-            )
-        })?;
-        self.ll.drx_tune1a().write(|w| {
-            w.value(
-                config
-                    .pulse_repetition_frequency
-                    .get_recommended_drx_tune1a(),
-            )
-        })?;
-        let drx_tune1b = config
-            .expected_preamble_length
-            .get_recommended_drx_tune1b(config.bitrate)?;
-        self.ll.drx_tune1b().write(|w| w.value(drx_tune1b))?;
-        let drx_tune2 = config
-            .pulse_repetition_frequency
-            .get_recommended_drx_tune2(
-                config.expected_preamble_length.get_recommended_pac_size(),
-            )?;
-        self.ll.drx_tune2().write(|w| w.value(drx_tune2))?;
-        self.ll
-            .drx_tune4h()
-            .write(|w| w.value(config.expected_preamble_length.get_recommended_dxr_tune4h()))?;
-
-        // Set channel tuning
-        self.ll
-            .rf_rxctrlh()
-            .write(|w| w.value(config.channel.get_recommended_rf_rxctrlh()))?;
-        self.ll
-            .fs_pllcfg()
-            .write(|w| w.value(config.channel.get_recommended_fs_pllcfg()))?;
-        self.ll
-            .fs_plltune()
-            .write(|w| w.value(config.channel.get_recommended_fs_plltune()))?;
-
-        // Set the LDE registers
-        self.ll
-            .lde_cfg2()
-            .write(|w| w.value(config.pulse_repetition_frequency.get_recommended_lde_cfg2()))?;
-        self.ll.lde_repc().write(|w| {
-            w.value(
-                config.channel.get_recommended_lde_repc_value(
-                    config.pulse_repetition_frequency,
-                    config.bitrate,
-                ),
-            )
-        })?;
-
-        // Check if the rx buffer pointer is correct
-        let status = self.ll.sys_status().read()?;
-        if status.hsrbp() != status.icrbp() {
-            // The RX Buffer Pointer of the host and the ic side don't point to the same one.
-            // We need to switch over
-            self.ll.sys_ctrl().modify(|_, w| w.hrbpt(1))?;
-        }
-
+    pub(super) fn start_receiving(&mut self) -> Result<(), Error<SPI>> {
         // Start receiving
         self.ll.sys_ctrl().modify(|_, w| w.rxenab(0b1))?;
 
@@ -318,6 +152,24 @@ where
             // without preventing a good frame from being received. Those are:
             // - LDEERR: Leading Edge Detection Processing Error
             // - RXPREJ: Receiver Preamble Rejection
+
+            if sys_status.txfrb() == 1
+                || sys_status.txprs() == 1
+                || sys_status.txphs() == 1
+                || sys_status.txfrs() == 1
+            {
+                // Old tx flags are still active. We don't need them anymore.
+                // This can happen from a tx continuation.
+                self.ll
+                    .sys_status()
+                    .write(|w| {
+                        w.txfrb(0b1) // Transmit Frame Begins
+                            .txprs(0b1) // Transmit Preamble Sent
+                            .txphs(0b1) // Transmit PHY Header Sent
+                            .txfrs(0b1) // Transmit Frame Sent
+                    })
+                    .map_err(|error| nb::Error::Other(Error::Spi(error.0)))?;
+            }
 
             // No errors detected. That must mean the frame is just not ready
             // yet.